Create bookmarks.ini (#103)
* Create bookmarks.ini

Add bookmark functionality

This allows you to set a mark on an item, then later restore that mark to the clipboard. 

This is similar to how vim registers work. 

The implementation uses special tags with a "mark:" prefix, and when a mark is set, removes  that tag from any items that contain that tag.

Can be used with

 itemtags\tags="\\\\1;;#0000ff;;\xe0bb;;;;mark:(.*);;"

to get nicer presentation of the marks.

* Add command for restoring bookmark